refactor: Move OpenAPI schema classes into a separate module
