+++ test/test_schemas.py
@@ -6,7 +6,7 @@ from jsonschema import ValidationError
 
 import schemathesis
 from schemathesis.exceptions import InvalidSchema
-from schemathesis.schemas import ConvertingResolver
+from schemathesis.specs.openapi.schemas import ConvertingResolver
 
 
 @pytest.mark.parametrize("base_path", ("/v1", "/v1/"))
@@ -46,7 +46,7 @@ def test_open_api_verbose_name(openapi_30):
 
 def test_resolver_cache(simple_schema, mocker):
     schema = schemathesis.from_dict(simple_schema)
-    spy = mocker.patch("schemathesis.schemas.ConvertingResolver", wraps=ConvertingResolver)
+    spy = mocker.patch("schemathesis.specs.openapi.schemas.ConvertingResolver", wraps=ConvertingResolver)
     assert "_resolver" not in schema.__dict__
     assert isinstance(schema.resolver, ConvertingResolver)
     assert spy.call_count == 1
